Add testing Calculate and Operate
